services/business-logic/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
